Use int64 type

+++ convert/convert.go
@@ -1,25 +1,25 @@
 package convert
 
 // pow10
-var pow10 = [...]int{
+var pow10 = [...]int64{
 	1e00, 1e01, 1e02, 1e03, 1e04, 1e05, 1e06, 1e07, 1e08, 1e09,
 	1e10, 1e11, 1e12, 1e13, 1e14, 1e15, 1e16, 1e17, 1e18,
 }
 
 // pow16
-var pow16 = [...]int{
+var pow16 = [...]int64{
 	1, 16, 256, 4096, 65536,
 	1048576, 16777216, 268435456, 4294967296, 68719476736,
 	1099511627776, 17592186044416, 281474976710656, 4503599627370496,
 }
 
-// DecToInt converts an ascii encoded decimal integer to an int.
-func DecToInt(slice []byte) int {
-	num := 0
+// DecToInt converts an ascii encoded decimal integer to an int64.
+func DecToInt(slice []byte) int64 {
+	num := int64(0)
 	length := len(slice)
 
 	for i := 0; i < length; i++ {
-		code := int(slice[i])
+		code := int64(slice[i])
 
 		if code >= 48 && code <= 57 {
 			num += (code - 48) * pow10[length-i-1]
@@ -29,13 +29,13 @@ func DecToInt(slice []byte) int {
 	return num
 }
 
-// HexToInt converts an ascii encoded hexadecimal integer to an int.
-func HexToInt(slice []byte) int {
-	num := 0
+// HexToInt converts an ascii encoded hexadecimal integer to an int64.
+func HexToInt(slice []byte) int64 {
+	num := int64(0)
 	length := len(slice)
 
 	for i := 0; i < length; i++ {
-		code := int(slice[i])
+		code := int64(slice[i])
 		pos := length - i - 1
 
 		switch {
